Resolve a bunch of `Multiplication result converted to larger type` finding (#760)
* Resolve several CodeQL findings

Specifically `Multiplication result converted to larger type`

+++ src/input.cc
@@ -386,7 +386,7 @@ void takeScreenshot()
 {
     int width = _scr_size.right - _scr_size.left + 1;
     int height = _scr_size.bottom - _scr_size.top + 1;
-    gScreenshotBuffer = (unsigned char*)internal_malloc(width * height);
+    gScreenshotBuffer = (unsigned char*)internal_malloc(static_cast<size_t>(width) * height);
     if (gScreenshotBuffer == nullptr) {
         return;
     }
